Alrighty, to preview the scene on a device, you must have the JDK and Android SDK. The Phone must be set to enable USB debugging.
Card: This defines our real world playing cards, they have a suit and a rank as enums for easy comparison for scoring/hand comparison.
CardHands: This defines as enums, our winning hands for different variants of poker. Could be extended for other card games?
CardTypes: This currently in the Poker Namespace divides cards into 4 suits and 13 ranks.
GameController: So far this handles the majority of our game logic. I believe for performance reasons we may wish to rely on message passing and event handling for progressing the state of the game, using the GameController as our central game state manager. The game table has a current pot, a current pool of shared cards, and a number of players.
Player: For AI or Humans, the player has a hand of cards, a 'wallet' or whatever the term is for your bankroll.

Today time was spent reviewing the rules and gameplay of poker, and iterating over the design of the title's backend.