Today's (Wednesday) Status Report: I decided that I will likely extremely soon hit a brick wall in terms of easily testable functionality so I switched focus to going over the Photon Networking for Unity3d tutorials.
I am pleased to report that confusing labeling aside you can hot join over the network with a build directly installed on your phone and on the Windows computer (Hitting the play button that is, I didn't try it with a Windows build); so we don't need any expensive licenses for simple 1v1 testing. I am going through the process of understanding synchronization and RPC calls via Photon and expect to be done sometime Thursday in order to progress to proceeding on implementing basic functionality.